# Faster chromatic symmetric function computation