--- a/src/sage/graphs/graph.py
+++ b/src/sage/graphs/graph.py
@@ -3930,14 +3930,13 @@ class Graph(GenericGraph):
 
         ALGORITHM:
 
-            We traverse a binary tree whose leaves correspond to
-            subsets of edges, and whose internal vertices at depth `d`
-            correspond to a choice of whether to include the `d`-th
-            edge in a given subset. The components of the induced
-            subgraph are incrementally updated with a disjoint-set
-            forest. If the next edge would introduce a cycle to the
-            subset, we prune the branch as the terms produced by the
-            two subtrees cancel in this case.
+        We traverse a binary tree whose leaves correspond to subsets of
+        edges, and whose internal vertices at depth `d` correspond to a
+        choice of whether to include the `d`-th edge in a given subset.
+        The components of the induced subgraph are incrementally
+        updated using a disjoint-set forest. If the next edge would
+        introduce a cycle to the subset, we prune the branch as the
+        terms produced by the two subtrees cancel in this case.
 
         EXAMPLES::