@mantepse 2024-03-09
> The algorithm traverses a binary tree whose leaves represent subsets of edges. The induced set partition on vertices is incrementally updated using a disjoint-set forest so that the set partition does not need to be completely recomputed for each subset. This approach further allows us to prune branches of the binary tree when the next edge of the graph would introduce a cycle to the subset, since the terms produced by the two subtrees cancel in this case.

I think it would be good to include this description (or something similar) in the code, either as `ALGORITHM:` after `INPUT:`, or simply as a comment in the code.