@tscrim 2024-03-10
Ah right. I forgot it no longer returns a list.